+++ ReleaseNotes.md
@@ -4,7 +4,7 @@
 * Add Output Window pane named "SwitchStartupProject"
 * Replace ActivityLog messages and popup messages with log messages in Output Window (issue #60)
 * Rename command and tooltip of dropdown to "SwitchStartupProject" to avoid confusion (issue #67)
-* To improve Visual Studio responsiveness load the SwitchStartupProject extension in the background
+* To improve Visual Studio responsiveness load the SwitchStartupProject extension in the background when a solution is opened
 
 ## Version 3.5
 * Fix support for configuration parameters of C++ projects (issues #55, #61)